Work on making routing domain configurable; more work on solvers-service
