+++ services/solvers-service/helm/Chart.yaml
@@ -1,5 +1,5 @@
 apiVersion: v2
-name: mzn-controller
+name: solvers-service
 description: MiniZinc solver controller
 
 # A chart can be either an 'application' or a 'library' chart.
