--- services/solvers-service/helm/templates/pod.yml
@@ -1,11 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: {{.Chart.Name}}
-  namespace: {{.Release.Namespace | quote}}
-spec:
-  serviceAccountName: "{{ .Chart.Name }}-serviceaccount"
-  containers:
-    - name: {{.Chart.Name}}
-      image: "{{.Values.image.repository}}:{{.Values.image.tag}}"
-
