+++ services/solvers-service/helm/templates/service-account.yml
@@ -9,7 +9,7 @@ apiVersion: rbac.authorization.k8s.io/v1
 kind: Role
 metadata:
   namespace: {{ .Release.Namespace | quote }}
-  name: mzn-solver-creator
+  name: solvers-service-creator
 rules:
 - apiGroups: [""]
   resources: ["pods", "pods/log"]
@@ -21,12 +21,12 @@ rules:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: RoleBinding
 metadata:
-  name: mzn-solver-creator
+  name: solvers-service-creator
   namespace: {{ .Release.Namespace | quote }}
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: Role
-  name: mzn-solver-creator
+  name: solvers-service-creator
 subjects:
 - kind: ServiceAccount
   name: "{{ .Chart.Name }}-serviceaccount"
