+++ services/solvers-service/helm/templates/service.yml
@@ -1,0 +1,12 @@
+apiVersion: serving.knative.dev/v1
+kind: Service
+metadata:
+  name: {{.Chart.Name}}
+  namespace: {{.Release.Namespace | quote}}
+spec:
+  template:
+    spec:
+      containers:
+      - image: "{{.Values.image.repository}}:{{.Values.image.tag}}"
+        ports:
+        - containerPort: 8080
