+++ services/solvers-service/package.json
@@ -1,11 +1,12 @@
 {
-  "name": "mzn-controller",
+  "name": "solvers-service",
   "version": "1.0.0",
   "main": "index.js",
   "type": "module",
   "dependencies": {
     "@kubernetes/client-node": "^0.17.1",
     "amqplib": "^0.10.3",
+    "express": "^4.18.2",
     "typescript": "^4.8.4"
   },
   "scripts": {
@@ -19,6 +20,7 @@
   "description": "",
   "devDependencies": {
     "@types/amqplib": "^0.8.2",
+    "@types/express": "^4.17.14",
     "@types/node": "^18.11.3",
     "nodemon": "^2.0.20",
     "prettier": "^2.7.1",
