+++ services/solvers-service/src/api/modelsRoute.ts
@@ -1,0 +1,11 @@
+import express from 'express'
+
+export default () => {
+  const jobs = express.Router()
+
+  jobs.get('*', (req, res) => {
+    res.send(`Solvers service /models: ${req.path}`)
+  })
+
+  return jobs
+}
