+++ services/solvers-service/src/api/solversRoute.ts
@@ -1,0 +1,11 @@
+import express from 'express'
+
+export default () => {
+  const jobs = express.Router()
+
+  jobs.get('*', (req, res) => {
+    res.send(`Solvers service /solvers: ${req.path}`)
+  })
+
+  return jobs
+}
